Moves ClockTimer out of logic.
