+++ src/gui/events/handlers/PauseGameEventHandler.java
@@ -4,6 +4,7 @@ import gui.events.IEventSubscriber;
 import gui.events.PauseGameEvent;
 import gui.events.SetResetButtonIconEvent;
 import gui.Resource;
+import gui.ClockTimer;
 import logic.game.*;
 
 public class PauseGameEventHandler implements IEventHandler<PauseGameEvent> {
